Updated find events screen
Picker is styled more similar to the figma prototype and will be constant across all platforms. Did not use built in picker library (since it was displaying differently on ios) so you must "npm install react-native-dropdown-picker" (external library) for it to work. Styling with the top navigation bar might be kind uneven so lmk if it looks like garbage. I haven't actually done anything for the search pop-up style-wise and I was going to finish that up later when we do the backend for it but lmk if you want it done now.